Fixed: Save and restore last stated after reset
It has now been tested and confirmed working!

+++ EEPROMAnything.h
@@ -8,18 +8,18 @@
 // Also added EEPROM_updateAnything to limit number of write/erase cycles
 // By: Krisitan Lauszus, TKJ Electronics
 
-template <class T> uint8_t EEPROM_writeAnything(uint8_t addr, const T& value) {
+template <class T> uint16_t EEPROM_writeAnything(uint16_t addr, const T& value) {
   eeprom_busy_wait(); // Wait until the eeprom is ready
   const uint8_t *p = (const uint8_t*)(const void*)&value;
-  uint8_t i;
+  uint16_t i;
   for (i = 0; i < sizeof(value); i++)
     eeprom_write_byte((uint8_t*)addr++, *p++);
   return i;
 }
-template <class T> uint8_t EEPROM_updateAnything(uint8_t addr, const T& value) {
+template <class T> uint16_t EEPROM_updateAnything(uint16_t addr, const T& value) {
   eeprom_busy_wait(); // Wait until the eeprom is ready
   const uint8_t *p = (const uint8_t*)(const void*)&value;
-  uint8_t i;
+  uint16_t i;
   for (i = 0; i < sizeof(value); i++) {
     if (eeprom_read_byte((uint8_t*)addr) != *p) // Limits number of write/erase cycles
       eeprom_write_byte((uint8_t*)addr, *p);
@@ -28,10 +28,10 @@ template <class T> uint8_t EEPROM_updateAnything(uint8_t addr, const T& value) {
   }
   return i;
 }
-template <class T> uint8_t EEPROM_readAnything(uint8_t addr, T& value) {
+template <class T> uint16_t EEPROM_readAnything(uint16_t addr, T& value) {
   eeprom_busy_wait(); // Wait until the eeprom is ready
   uint8_t *p = (uint8_t*)(void*)&value;
-  uint8_t i;
+  uint16_t i;
   for (i = 0; i < sizeof(value); i++)
     *p++ = eeprom_read_byte((uint8_t*)addr++);
   return i;
